Move Hash to package `common`; use Hash in Hashifier
