+++ src/main/java/jp/co/soramitsu/sora/crypto/common/Hash.java
@@ -1,4 +1,4 @@
-package jp.co.soramitsu.sora.crypto.merkle;
+package jp.co.soramitsu.sora.crypto.common;
 
 import javax.xml.bind.DatatypeConverter;
 import lombok.Value;
