+++ src/main/java/jp/co/soramitsu/sora/crypto/common/Util.java
@@ -6,7 +6,6 @@ import java.security.MessageDigest;
 import java.security.Signature;
 import java.security.SignatureException;
 import jp.co.soramitsu.sora.crypto.json.JSONCanonizer;
-import jp.co.soramitsu.sora.crypto.merkle.Hash;
 import jp.co.soramitsu.sora.crypto.proof.Options;
 import org.spongycastle.util.Arrays;
