+++ src/main/java/jp/co/soramitsu/sora/crypto/merkle/MerkleNode.java
@@ -1,5 +1,6 @@
 package jp.co.soramitsu.sora.crypto.merkle;
 
+import jp.co.soramitsu.sora.crypto.common.Hash;
 import lombok.Value;
 
 @Value
