+++ src/main/java/jp/co/soramitsu/sora/crypto/merkle/MerkleTreeProof.java
@@ -6,6 +6,7 @@ import java.security.MessageDigest;
 import java.util.Comparator;
 import java.util.LinkedList;
 import java.util.List;
+import jp.co.soramitsu.sora.crypto.common.Hash;
 import jp.co.soramitsu.sora.crypto.common.Util;
 import lombok.NonNull;
 import lombok.Value;
