+++ src/test/groovy/jp/co/soramitsu/sora/crypto/MerkleTreeTest.groovy
@@ -3,6 +3,7 @@ package jp.co.soramitsu.sora.crypto
 import jp.co.soramitsu.sora.common.MockSumMessageDigest
 import jp.co.soramitsu.sora.crypto.common.ArrayTree
 import jp.co.soramitsu.sora.crypto.common.ArrayTreeFactory
+import jp.co.soramitsu.sora.crypto.common.Hash
 import jp.co.soramitsu.sora.crypto.merkle.*
 import spock.lang.Specification
 import spock.lang.Unroll
